requirements.txt
fastapi==0.86.0
uvicorn== 0.19.0
pydantic==1.10.2
sqlalchemy==1.4.43
sqlalchemy_utils==0.38.3
python-multipart==0.0.5
aiofiles==22.1.0
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
fastapi_jwt_auth==0.5.0
werkzeug==2.2.2
pytz==2022.6
uuid==1.30
aiofiles==22.1.0
requests==2.28.1
pyttsx3==2.90
mysql-connector-python==8.0.31
thermalprinter==0.2.0
Authlib==1.2.1
redis==4.6.0
python-dotenv==1.0.0

